Fixed the speed display
git-svn-id: file:///Users/lightster/backup-8.31.2013_20-06-44_litesign/homedir/svn/Workspaces/Dev@50 3baa19a1-490e-4327-b14f-f43bf3c759de

+++ nascar/src/com/lightdatasys/nascar/fantasy/gui/ResultCell.java
@@ -128,7 +128,7 @@ public class ResultCell extends Cell
 		}
 		else if(mode == Mode.SPEED)
 		{
-			return String.format("%.1f", result.getSpeed());
+			return String.format("%d", (int)result.getSpeed());
 		}
 		
 		return (new Integer(result.getFinish())).toString();
